# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Rese",
        "description": "Rese API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server (php artisan serve)"
        },
        {
            "url": "http://rese-backend.test",
            "description": "Local server (Herd)"
        }
    ],
    "paths": {
...
        "/customers/{customer}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "顧客情報取得",
                "description": "セッション中の顧客情報を取得する",
                "operationId": "get-customer",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/customer-id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/show-customer-200"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/not-found"
                    }
                }
            }
        },
        "/customers/{customer}/shops/{shop}/favorite": {
            "post": {
                "tags": [
                    "Favorite"
                ],
                "summary": "お気に入り登録",
                "description": "セッション中の顧客が指定の飲食店をお気に入り登録する",
                "operationId": "post-customer-shop-favorite",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/customer-id"
                    },
                    {
                        "$ref": "#/components/parameters/shop-id"
                    }
                ],
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/created"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/not-found"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessable-entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Favorite"
                ],
                "summary": "お気に入り解除",
                "description": "セッション中の顧客が指定の飲食店のお気に入りを解除する",
                "operationId": "delete-customer-shop-favorite",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/customer-id"
                    },
                    {
                        "$ref": "#/components/parameters/shop-id"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/no-content"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/not-found"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessable-entity"
                    }
                }
            }
        },
        "/customers/{customer}/shops/{shop}/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "飲食店詳細ページの予約一覧取得",
                "description": "セッション中の顧客が指定の飲食店で行っている予約を一覧取得する",
                "operationId": "get-customer-shop-reservations",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/customer-id"
                    },
                    {
                        "$ref": "#/components/parameters/shop-id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/get-customer-shop-reservations-200"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/not-found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "飲食店詳細ページの予約追加",
                "description": "セッション中の顧客が指定の飲食店で予約を追加する",
                "operationId": "post-customer-shop-reservations",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/customer-id"
                    },
                    {
                        "$ref": "#/components/parameters/shop-id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/post-customer-shop-reservations"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/post-customer-shop-reservations-201"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/not-found"
                    },
                    "422": {
                        "$ref": "#/components/responses/post-customer-shop-reservations-422"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "area-data": {
                "description": "エリア情報",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "エリアID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "エリア名",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "genre-data": {
                "description": "ジャンル情報",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ジャンルID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "ジャンル名",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "shop-data": {
                "description": "飲食店情報",
                "required": [
                    "id",
                    "name",
                    "area",
                    "genre",
                    "image_url",
                    "detail",
                    "favorite_status"
                ],
                "properties": {
                    "id": {
                        "description": "飲食店ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "飲食店名",
                        "type": "string"
                    },
                    "area": {
                        "$ref": "#/components/schemas/area-data"
                    },
                    "genre": {
                        "$ref": "#/components/schemas/genre-data"
                    },
                    "image_url": {
                        "description": "画像URL",
                        "type": "string",
                        "format": "uri"
                    },
                    "detail": {
                        "description": "飲食店詳細",
                        "type": "string"
                    },
                    "favorite_status": {
                        "description": "お気に入りステータス",
                        "type": "string",
                        "enum": [
                            "unknown",
                            "marked",
                            "unmarked"
                        ]
                    }
                },
                "type": "object"
            },
            "reservation-data": {
                "description": "予約情報",
                "required": [
                    "id",
                    "shop",
                    "reserved_at",
                    "number_of_guests"
                ],
                "properties": {
                    "id": {
                        "description": "予約ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "shop": {
                        "$ref": "#/components/schemas/shop-data"
                    },
                    "reserved_at": {
                        "description": "予約日時",
                        "type": "string",
                        "format": "date-time"
                    },
                    "number_of_guests": {
                        "description": "予約人数",
                        "type": "integer",
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "reservation-error": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "properties": {
                            "reserved_at": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "number_of_guests": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "show-customer-200": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "テストユーザー"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "created": {
                "description": "リクエストが成功しリソースが作成された"
            },
            "no-content": {
                "description": "リクエストが成功しリソースが存在しない"
            },
...
            "unauthorized": {
                "description": "認証エラー"
            },
            "forbidden": {
                "description": "認可エラー"
            },
            "not-found": {
                "description": "指定されたリソースが存在しない"
            },
...
            "unprocessable-entity": {
                "description": "リクエストが正しくない"
            },
...
            "get-customer-shop-reservations-200": {
                "description": "飲食店詳細ページの予約一覧取得成功",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "reservations"
                            ],
                            "properties": {
                                "reservations": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/reservation-data"
                                    }
                                }
                            },
                            "type": "object"
                        },
                        "examples": {
                            "example": {
                                "$ref": "#/components/examples/get-customer-shop-reservations-200-example"
                            },
                            "empty": {
                                "$ref": "#/components/examples/get-customer-shop-reservations-200-empty"
                            }
                        }
                    }
                }
            },
            "post-customer-shop-reservations-201": {
                "description": "飲食店詳細ページの予約追加成功",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "reservation"
                            ],
                            "properties": {
                                "reservation": {
                                    "$ref": "#/components/schemas/reservation-data"
                                }
                            },
                            "type": "object"
                        },
                        "examples": {
                            "example": {
                                "$ref": "#/components/examples/post-customer-shop-reservations-201"
                            }
                        }
                    }
                }
            },
            "post-customer-shop-reservations-422": {
                "description": "飲食店詳細ページの予約追加のバリデーションエラー",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/reservation-error"
                        },
                        "examples": {
                            "required": {
                                "$ref": "#/components/examples/post-customer-shop-reservations-422-required"
                            },
                            "date": {
                                "$ref": "#/components/examples/post-customer-shop-reservations-422-reserved_at"
                            },
                            "past": {
                                "$ref": "#/components/examples/post-customer-shop-reservations-422-past"
                            },
                            "number_of_guests": {
                                "$ref": "#/components/examples/post-customer-shop-reservations-422-number_of_guests"
                            }
                        }
                    }
                }
            }
        },
        "parameters": {
            "customer-id": {
                "name": "customer",
                "in": "path",
                "description": "顧客ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                }
            },
            "shop-id": {
                "name": "shop",
                "in": "path",
                "description": "飲食店ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                }
            },
...
        },
        "examples": {
...
            "get-customer-shop-reservations-200-example": {
                "summary": "予約一覧の例",
                "value": {
                    "reservations": [
                        {
                            "id": 1,
                            "shop": {
                                "id": 1,
                                "name": "店舗1",
                                "area": {
                                    "id": 1,
                                    "name": "東京都"
                                },
                                "genre": {
                                    "id": 1,
                                    "name": "和食"
                                },
                                "image_url": "https://api.rese.com/images/1.jpg",
                                "detail": "サンプルテキスト",
                                "favorite_status": "unmarked"
                            },
                            "reserved_at": "2021-12-31T18:00:00+09:00",
                            "number_of_guests": 2
                        },
                        {
                            "id": 2,
                            "shop": {
                                "id": 2,
                                "name": "店舗2",
                                "area": {
                                    "id": 2,
                                    "name": "大阪府"
                                },
                                "genre": {
                                    "id": 2,
                                    "name": "中華"
                                },
                                "image_url": "https://api.rese.com/images/2.jpg",
                                "detail": "サンプルテキスト",
                                "favorite_status": "marked"
                            },
                            "reserved_at": "2022-01-01T19:00:00+09:00",
                            "number_of_guests": 4
                        }
                    ]
                }
            },
            "get-customer-shop-reservations-200-empty": {
                "summary": "該当する予約がない場合",
                "value": {
                    "reservations": []
                }
            },
            "post-customer-shop-reservations-201": {
                "summary": "予約追加成功",
                "value": {
                    "id": 1,
                    "shop": {
                        "id": 1,
                        "name": "店舗1",
                        "area": {
                            "id": 1,
                            "name": "東京都"
                        },
                        "genre": {
                            "id": 1,
                            "name": "和食"
                        },
                        "image_url": "https://api.rese.com/images/1.jpg",
                        "detail": "サンプルテキスト",
                        "favorite_status": "marked"
                    },
                    "reserved_at": "2022-01-01T19:00:00+09:00",
                    "number_of_guests": 4
                }
            },
            "post-customer-shop-reservations-422-required": {
                "summary": "予約日時、人数が入力されていない場合",
                "value": {
                    "message": "予約日時を入力してください (and 1 more error)",
                    "errors": {
                        "reserved_at": [
                            "予約日時を入力してください"
                        ],
                        "number_of_guests": [
                            "人数を入力してください"
                        ]
                    }
                }
            },
            "post-customer-shop-reservations-422-reserved_at": {
                "summary": "予約日時がRFC3339形式でない場合",
                "value": {
                    "message": "予約日時はRFC3339形式で入力してください",
                    "errors": {
                        "reserved_at": [
                            "予約日時はRFC3339形式で入力してください"
                        ]
                    }
                }
            },
            "post-customer-shop-reservations-422-past": {
                "summary": "過去の日時を指定した場合",
                "value": {
                    "message": "過去の日時は指定できません",
                    "errors": {
                        "reserved_at": [
                            "過去の日時は指定できません"
                        ]
                    }
                }
            },
            "post-customer-shop-reservations-422-number_of_guests": {
                "summary": "人数が0以下の場合",
                "value": {
                    "message": "人数は1以上で入力してください",
                    "errors": {
                        "number_of_guests": [
                            "人数は1以上で入力してください"
                        ]
                    }
                }
            }
        },
        "requestBodies": {
...
            "post-customer-shop-reservations": {
                "description": "予約追加リクエスト",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "reserved_at",
                                "number_of_guests"
                            ],
                            "properties": {
                                "reserved_at": {
                                    "description": "予約日時",
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2021-12-31T23:59:59+09:00"
                                },
                                "number_of_guests": {
                                    "description": "予約人数",
                                    "type": "integer",
                                    "minimum": 1,
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
}